moved data processing code in App.js to backend server and commented out the api request code on category filter
