+++ news-around-app/src/App.js
@@ -2,88 +2,39 @@
 import './App.css';
 import {Box, Image, HStack, Heading} from '@chakra-ui/react'
 import * as React from 'react'
-import {useState, useEffect} from 'react'
+import { useEffect } from 'react'
 import CountryFilter from './Components/CountryFilter';
 import CategoryFilter from './Components/CategoryFilter';
 import NewsArticle from './Components/NewsArticle';
 
 function App() {
-  // this part is for the newsHeadline api endoint
-  // still need to find a way to send language to the backend from the frontend
-  // const [ language, setLanguage ] = useState(null);
- 
-  const [allArticles, setAllArticles] = useState([]); // set allArticles as an empty array by default
 
-  const getHeadlines = async () => {
-    try {
-      const language = 'en'
-      const response = await fetch(`http://localhost:8000/headlines?language=${language}`);
-      // save all data from from backend server retrieved from News API
-      const data = await response.json();
-      console.log(data); // handle the response from the backend
-      // save all the article objects into articleData
-      const articleData = data.articles;
-      // use the setAllArticle function to save each article object in allArticles
-      setAllArticles(Object.values(articleData));
-    } catch (error) {
-      console.error(error);
-    }
-  };
+  // Save user's preferred browser language in usrlang
+  let usrlang = navigator.languages[0];
+  if (usrlang.includes("-")) {
+    usrlang = usrlang.split("-")[0]
+  }
+  console.log(usrlang); 
+
+  // this part is for the newsHeadline api endoint 
+  let data = [] // this allows data from the headlines endpoint to be stored globally
 
   useEffect(() => {
+    const getHeadlines = async () => {
+      try {
+        const language = usrlang;
+        const response = await fetch(`http://localhost:8000/headlines?language=${language}`);
+        const data = await response.json();
+        console.log(data);
+      } catch (error) {
+        console.error(error);
+      }
+    };
+  
     getHeadlines();
-  },[]); // use the useEffect hook to call the getHeadlines function when App renders
+  }, [usrlang]);
 
-
-  // use a for loop to print out each articles title/source/author etc
-  // checked description, urlToImage and content properties, does not retrieve anything
-
-  // store each article's data in an array where each article is stored in a dictionary
-  const allArticleData = []
-
-    for (let i = 0; i < allArticles.length; i++) {
-      let hyphenAmount = allArticles[i].title.split("-").length -1
-      let articleTitle = allArticles[i].title.split("-")[0]
-      let sourceCheck = ""
-      let author = allArticles[i].author
-      let newsSource = allArticles[i].source["Name"]
-      let url = allArticles[i].url
-      // clean publishDate property
-      let publishDate = allArticles[i].publishedAt.split("T")[0]      
-      // check title property for more than one hyphens
-      if (hyphenAmount > 1) {
-        // console.log("theres more than 1 hyphen")
-        sourceCheck = allArticles[i].title.split("-")[hyphenAmount]
-        articleTitle = allArticles[i].title.replace(sourceCheck, '')
-        // get the lastIndex of "-"
-        let lastHyphenIndex = articleTitle.lastIndexOf("-")
-        // use subString method to only save only up to the lastHyphen without the hyphen itself
-        articleTitle = articleTitle.substring(0, lastHyphenIndex)
-      }
-      // some article titles include | News Source or News Type, code below to get rid of it
-      if (articleTitle.includes("|")) {
-        articleTitle = articleTitle.split("|")[0]
-      }
-      // check author property
-      // clean up author when it is null or an empty string 
-      if (author === null || author === "") {
-        author = "N/A"
-      }
-      let articleData = {
-        title: articleTitle,
-        author: author,
-        source: newsSource, 
-        url: url, 
-        date: publishDate // 2023-05-16
-      };
-
-      allArticleData.push(articleData);
-  }
-
-  // check original article data retrieved from headlines endpoint
-  // console.log(allArticles)
-  // check article data after data clean up on headlines endpoint
-  // console.log({allArticleData})
+  console.log(data)
 
   return (
     <>
@@ -98,7 +49,7 @@ function App() {
         <CategoryFilter/>
         <CountryFilter />
         </HStack>
-        <NewsArticle title={allArticleData[0]?.title} author={allArticleData[0]?.author}/>
+        {/* <NewsArticle title={allArticleData[0]?.title} author={allArticleData[0]?.author}/> */}
 
 </>
   );
